Configure project to be published from build dir
This is apparently necessary to ensure that you can properly import
submodules (`import * as Arrays from 'potence/arrays') as you'd have to
add the build folder to the import path otherwise.

Also, added a prepublish script that causes the publish process to fail
if it is run from the root directory instead of the build directory.