Use object in place of ObjectLiteral for arguments
One big problem consumers of potence faced is that many functions
expected an ObjectLiteral as one of its arguments. As a consequence,
this meant that you could not simply pass any object in, even an object
literal which was typed by an interface, because the object's type
lacked an index signature.

This has now been remedied.