Change <p> to <div>
This fixes a problem where attempting to use the markdownify filter
inside of a <p> tag will put the markdownified text *outside* the <p>
tag, in a new <p> tag. This is obviously not what is wanted here and
resulted in an empty <p class="description"> tag, which meant that the
description margins were messed up.