Remove Strings.match()
Benchmarking showed that Strings.match() was over 90% slower than
corresponding RegExes on the same strings. Even a complete rework of the
algorithm to minimize string allocations and other inefficiencies still
resulted in a function that was around 30% slower than the RegExp.

Since I am sure that, aside from minor negligible optimizations, the new
function (not committed) could not have been improved further, I am led
to the unfortunate conclusion that the kind of performance I was looking
for is not possible in plain JavaScript. The RegExp engine, of course,
is not written in JS code but is implemented by the runtime environment,
usually in C++, which will therefore be blazingly fast.

Even something as simple as iterating through a string and constructing
a new string by adding each char in sequence to the new string was
incredibly slow. In any other language iterating through a string and
conducting complex operations on the characters within would have been
faster by at least 8000%.

Since the main reason for the Strings.match() function was performance
over RegExes, which is simply not possible to achieve in JavaScript, I
have removed the function and all related types.