Disable Flags throwing on 0 flag
Throwing on a 0 flag doesn't make much sense. Of course, passing
a 0 flag is also usually illogical, but it can happen, especially
when one is passing a variable with a calculated value which may or
may not be 0. Thus, all operations on 0 flags will now succeed, they
will simply have no effect.